backend/data_collectors/requirements_kiwoom.txt
pykiwoom>=1.1.25
PyQt5>=5.15.7
pandas>=1.5.3
numpy>=1.24.3
supabase>=2.0.0
python-dotenv>=1.0.0